fix: rename contentresource to websitepageresource
