--- app/Filament/Admin/Resources/ContentsResource/Pages/ViewContents.php
@@ -1,12 +1,0 @@
-<?php
-
-namespace App\Filament\Admin\Resources\ContentsResource\Pages;
-
-use App\Filament\Admin\Resources\ContentsResource;
-use Filament\Actions;
-use Filament\Resources\Pages\ViewRecord;
-
-class ViewContents extends ViewRecord
-{
-    protected static string $resource = ContentsResource::class;
-}
