+++ app/Filament/Admin/Resources/WebsitePageResource/Api/Handlers/CreateHandler.php
@@ -1,13 +1,13 @@
 <?php
-namespace App\Filament\Admin\Resources\ContentsResource\Api\Handlers;
+namespace App\Filament\Admin\Resources\WebsitePageResource\Api\Handlers;
 
 use Illuminate\Http\Request;
 use Rupadana\ApiService\Http\Handlers;
-use App\Filament\Admin\Resources\ContentsResource;
+use App\Filament\Admin\Resources\WebsitePageResource;
 
 class CreateHandler extends Handlers {
     public static string | null $uri = '/';
-    public static string | null $resource = ContentsResource::class;
+    public static string | null $resource = WebsitePageResource::class;
 
     public static function getMethod()
     {
