+++ app/Filament/Admin/Resources/WebsitePageResource/Pages/EditWebsitePage.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace App\Filament\Admin\Resources\WebsitePageResource\Pages;
+
+use App\Filament\Admin\Resources\WebsitePageResource;
+use Filament\Resources\Pages\EditRecord;
+
+class EditWebsitePage extends EditRecord
+{
+    protected static string $resource = WebsitePageResource::class;
+
+    protected function getRedirectUrl(): string
+    {
+        return $this->getResource()::getUrl('view', ['record' => $this->record]);
+    }
+}
