+++ app/Filament/Admin/Resources/WebsitePageResource/Pages/ViewWebsitePage.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace App\Filament\Admin\Resources\WebsitePageResource\Pages;
+
+use App\Filament\Admin\Resources\WebsitePageResource;
+use Filament\Actions;
+use Filament\Resources\Pages\ViewRecord;
+
+class ViewWebsitePage extends ViewRecord
+{
+    protected static string $resource = WebsitePageResource::class;
+}
